+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+* v0.0.8 / 2020-08-28
+  * 修正vcheck.str，過濾不可見字元
+  * 修正vcheck.url，加入wildcard,protocols的參數
+  * 移除vutils.useragent、vutils.clientIP、vutils.replaceAll
+  * 新增vutils.arrayDistinct
+  * 移除mongo-sanitizer的dependency
+
 * v0.0.7 / 2019-09-11
   * 修正vcheck.url，加入允許localhost的參數